// apps/web/src/app/api/models/vehicle-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface VehicleDto {
  createdAt?: string;
  currentMileage?: number;
  id?: string;
  licensePlate?: string | null;
  make?: string | null;
  model?: string | null;
  updatedAt?: string;
  vin?: string | null;
  year?: number;
}